requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
sphinx 
sphinx-autodoc-typehints
alembic
python-jose
passlib[bcrypt]
"bcrypt>=4.0.0"